Rename sort to sorted and implement it
Implement sorting by splitting the pipeline each time a sort is requested.
At the end of each pipeline, the code checks if the pipeline needs sorting,
if it does, the values are collected to an array to allow sorting them using
usort.
It might be possible to enhance this because the call to iterator_to_array
forces the execution of the previous operations delayed by the yield usage